requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
sentence-transformers>=2.2.2
pydantic>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
torch>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0 